Moved blitTiling and BlitBatcher to ClientUtils
per code change request
